backend/requirements.txt
alembic
email-validator
fastapi
pwdlib[argon2]
PyJWT
psycopg[binary]
pydantic-settings
python-dotenv
SQLAlchemy
uvicorn[standard]
